% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{print.psychnet}
\alias{print.psychnet}
\title{Print a psychnet network}
\usage{
\method{print}{psychnet}(x, ...)
}
\arguments{
\item{x}{A \code{psychnet} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a psychnet network
}
\examples{
fit <- pcor_network(SRL_GPT)
print(fit)
}
