% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgm_moderated.R
\name{print.psychnet_moderated}
\alias{print.psychnet_moderated}
\title{Print a moderated MGM fit}
\usage{
\method{print}{psychnet_moderated}(x, ...)
}
\arguments{
\item{x}{A \code{psychnet_moderated} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a moderated MGM fit
}
