% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redundancy.R
\name{print.psychnet_redundancy}
\alias{print.psychnet_redundancy}
\title{Print a redundancy result}
\usage{
\method{print}{psychnet_redundancy}(x, ...)
}
\arguments{
\item{x}{A \code{psychnet_redundancy} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a redundancy result
}
